Interpolate between old and new IV samples instead of jumping

+++ core/src/libraries/constants/Constants.sol
@@ -123,7 +123,7 @@ uint32 constant IV_SCALE = 24 hours;
 /// @dev The initial value of implied volatility, used when `VolatilityOracle.prepare` is called for a new pool.
 /// Expressed as a 1e12 percentage at `IV_SCALE`, e.g. {0.12e12, 24 hours} → 12% daily → 229% annual. Error on the
 /// side of making this too large (resulting in low LTV).
-uint128 constant IV_COLD_START = 0.127921282726e12;
+uint104 constant IV_COLD_START = 0.127921282726e12;
 
 /// @dev The maximum rate at which (reported) implied volatility can increase. Raw samples in `VolatilityOracle.update`
 /// are clamped (before being stored) so as not to exceed this rate.
